mapeo casi terminado falta definir el pago

+++ src/main/java/com/example/ecommerce/Model/ItemOrden.java
@@ -6,15 +6,16 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ItemOrden {
+    
    private Producto producto;
-   private float precio;
+   private Double precio;
    private int cantidad;
-   public ItemOrden(Producto producto, float precio) {
+   public ItemOrden(Producto producto, Double precio) {
     this.producto = producto;
     this.precio = precio;
    }
 
-    public Float getSubtotal(){
+    public Double getSubtotal(){
         return  this.cantidad * producto.getPrecio();
     }
